tui: Tuxels no longer need Arc<Mutex<_>>
This is the culmination of recent commits which refactored the Canvas
Stack grid to deal with Cells rather than directly with Tuxels. Cells
provide a wrapper around owned Tuxels and Tuxels that can be accessed
via DrawBuffers aka DBTuxels.

This allows each Tuxel to be owned by either the Canvas OR a DrawBuffer
while using the DBTuxel abstraction to access the contents of Tuxel
owned by DrawBuffers (protected by the DrawBuffer's Arc<Mutex<_>>).